+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * STL: from stream (use BinaryFile and StringFile instead of io.BinaryIO and FileIO)
 * Step: from stream (use BinaryFile instead of io.BinaryIO)
 * Contour: is_overlapping (consider intersecting_points is empty)
+* LineSegment2D: to_wire (use discretization_points instead of discretise)
 
 
 
@@ -40,10 +41,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Avoid unneeded bbox computation
 
-
 ### Refactorings
 * LineSegment3D.intersections
-
 
 
 ### Unittests
@@ -61,7 +60,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * LineSegment3D: line_intersections
 * LineSegment3D: linesegment_intersections
 * Contour: is_overlapping
-
+* LineSegment2D: to_wire
 
 
 ## v0.6.0 [11/7/2022]